[Common] Refactor code
