+++ src/components/Header/mains/ContainerHeader/style.scss
@@ -5,11 +5,11 @@
   z-index: 20;
 }
 .btn-lang {
+  background-color: #fff;
   padding: 5px 20px;
   border: none;
   border-radius: 10px;
   margin: 0 10px;
-  background-color: #fff;
   font-size: 10px;
   &:active {
     background-color: #2ecc71;
