+++ src/contants/index.js
@@ -2,11 +2,9 @@ import PATH from "./PATH";
 import ROUTERS from "./routers/ROUTER";
 import ACTION_TYPE from "./ActionType/Home/actionType";
 import KEY_CODE from "./KeyCode/keyCode";
-import COLOR from "./Color/color";
 
 const CONSTANTS = {
   PATH,
-  COLOR,
   ROUTERS,
   ACTION_TYPE,
   KEY_CODE,
