+++ src/hooks/index.js
@@ -1,4 +1,4 @@
 export { default as useKeyPress } from "./useKeyPress";
 export { default as useHover } from "./useHover";
 export { default as useMultiLang } from "./useMultiLang";
-export { default as useMultiColor } from "./useMultiColor";
+export { default as useSetting } from "./useSetting";
