+++ src/pages/Home/components/ZingChartAlbumHeader/index.jsx
@@ -1,12 +1,12 @@
 // libs
 import React from "react";
 // hooks
-import { useMultiColor } from "../../../../hooks";
+import { useSetting } from "../../../../hooks";
 // orthers
 import "./style.scss";
 
 export default function ZingChartAlbumHeader() {
-  const { currentColor } = useMultiColor();
+  const { currentColor } = useSetting();
   return (
     <div className="zing-chart-album-header-wappper">
       <h1 style={{ color: currentColor }} className="zing-chart-album-title">
